rag/requirements.txt
flask
transformers
torch
sentence-transformers
faiss-cpu
openai
tiktoken
langchain
langchain-community
